[refactor] Change DIR
- 파일 구조 분리
related to #31
