+++ search/src/main/java/com/bjcareer/search/MongoDBConfig.java
@@ -1,0 +1,24 @@
+package com.bjcareer.search;
+
+import org.bson.Document;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import com.mongodb.client.MongoClients;
+import com.mongodb.client.MongoDatabase;
+
+@Configuration
+public class MongoDBConfig {
+
+	@Value("${spring.data.mongodb.uri}")
+	private String uri;
+
+	@Value("${spring.data.mongodb.database}")
+	private String database;
+
+	@Bean
+	public MongoDatabase mongoClient() {
+		return MongoClients.create(uri).getDatabase(database);
+	}
+}
