+++ search/src/main/java/com/bjcareer/search/retrieval/noSQL/MongoNode.java
@@ -1,6 +1,4 @@
-package com.bjcareer.search.retrieval;
-
-import java.util.Map;
+package com.bjcareer.search.retrieval.noSQL;
 
 public class MongoNode {
 	String keyword = "";
