+++ search/src/main/java/com/bjcareer/search/retrieval/noSQL/MongoQueryKeywords.java
@@ -1,9 +1,11 @@
-package com.bjcareer.search.retrieval;
+package com.bjcareer.search.retrieval.noSQL;
 
 public class MongoQueryKeywords {
 	public static final String KEYWORD = "keyword";
 	public static final String PARENT_ID = "parentId";
 	public static final String SEARCH_COUNT = "searchCount";
 	public static final String END_OF_WORD = "endOfWord";
+
+	public static final String CHILDS = "childs";
 	public static final String KEY = "_id";
 }
